add grpc geometry interface
